Added simple markdown renderer.
Added the beginnings of a markdown syntax highlighter, which doubles
as a git edit message highlighter.

Only lines starting with # are highlighted, and a *, +, or - character at the
start of a line followed by a space is assumed to be the start of a list, and
the one character is highlighted.