Adding an experimental interactive Shell plugin
This is mostly a proof-of-concept to test the extensibility
of the plugin system, but it adds a Shell command which
turns the active session into an interactive dumb terminal.

It's not as fancy as win is in ACME and still a work-in-progress, but
it seems to be working well enough without causing other bugs to commit.