Improved memory usage for large files.
The Renderer interface was changed to return an image, and a rectangle
denoting how large the image *would* be if the entire file was rendered.
(so that the viewport knows if it can scroll down or up or should prevent
the user from scrolling any further.)

This drastically reduces the memory usage, since now a renderer only needs
to return an image with the part of the image that's visible on the screen
rendered, instead of the entire file. (Returning an RGBA of the correct
size but without anything drawn wouldn't help memory usage much, since the
NewRGBA call would still need to allocate 4 bytes*X*Y, even if it doesn't
draw to it.)

It has the side-effect of making the renderer much more efficient, even for
small files, and the keyboard feels much more responsive.

The responsiveness of the keyboard becomes somewhat slow with large files,
since calculating the size of the rectangle involves going through the entire
file (but it had to do that before this change anyways.)