Cache image size calculation
The change to fix the memory consumption fixed the memory
consumption of large files, but the render time was still
a function of the file size since calculating the image size
required going through the whole buffer. The calculation is
now cached, so as long as the buffer size hasn't changed, and
the buffer hasn't changed, is just returns the previously
calculated image.Rectangle. This should result in browsing
large files being much more responsive. The only pathological
case that remains is editing large files.