Improved usage of external commands
When external commands are executed, they now get the selected text (or
whole buffer if nothing is selected) on STDIN, and can be used to
either replace the file, or update a selection (ie. by executing the
word |gofmt or |sed)

See USAGE.md for more details.

Fixes #7, Fixes #4 (or at least obsoletes #4.)