Revert usage of AltDot to make execution more reliable.
The AltDot seems to have been prematurely commited and wasn't as reliable
as I thought.

This reverts all usages of it to just use normal Dot, which results in
executing words being much more reliable. This means that for most things
that are run through the shell, it needs to be done through the tagline
since otherwise most commands would just receive themselves on STDIN.

There's still a bug where the last word in the tag line won't execute
(which makes :cmd<Enter> from the documentation a lie), but overall things
are more accurate now.