Actually use DPI from size.Event message
There were a couple stupid bugs in commit b12912bec. In particular:

1. The RecalculateFontFace wasn't *using* the DPI it was passed, and
   still assuming 72.0
2. The DPI wasn't being calculated properly. There's 72 points in an
   inch, so PixelsPerPoint needs to be multiplied by 72 to get DPI.

This should fix it. As a result, now that the font is being parsed
at the real DPI, the font size can be shrunk. Previously, it was using
at 16pt font because the font was parsed as 72 DPI but usually used on
a 96 DPI screen, effectively making it 12pt. With the font being parsed
at the screen resolution, we can change it to really be 12pt.