Added image viewer mode
Since syntax highlighting is done by rendering to an image and
golang's image package includes decoders, there's no reason we
can't check for the magic byte and render the actual image if
appropriate.

This is potentially useful for working with web apps, where you
might want to view the assets for your web app, but mostly it's
a proof of concept, because we can.