Invalidate size cache on DPI change.
The renderers calculate the initial size of the bounding box before
we know the screen resolution and cache the results. This causes the
bounds to be wrong, and scrolling to stop, until the buffer is changed.

This invalidates the cache as soon as we get a size event, so that things
work properly.