Fixed crash under Windows
The windows driver was (quite reasonably) returning an error
when trying to allocate the buffer in paintWindow, before the
screen has been initialized and sent a size event. The error
handling code wasn't handled properly, but other drivers don't
seem to return an error in this situation.

This fixes it by:
1. Returning, instead of panicing
2. Moving the defered release to after the err checking, as
   should be done in idiomatic Go anyways. (The call was
   throwing a nil pointer panic otherwise.)