Fixed bug where <n>G wasn't updating the viewport to the right line.
The render method was only calculating the imagemap for the portion
of the viewport that was being rendered to the screen, so when G tried
to use the imagemap to find the location of the line that it's going to,
it was having an error returned claiming there was no character in the
image map.

This separates the image map from the render code and caches the imagemap,
so that we always have a full image map. As a bonus, the render method
does less and everything feels more responsive.