Changed behaviour of Escape key to require less mouse usage
The escape key currently now Save if the buffer is dirty, and Exit
if it isn't. This means that you can easily Save by just hitting escape,
or save and exit by hitting it twice. It's less intuitive, but more useful.

The Exit command is also safer, in that it will only Quit if the buffer
is clean, otherwise provide a warning asking you to Save, Discard, or
ForceQuit

SaveAndExit is also an option now, as well as SaveOrExit and every
other combination I could think of about what you'd like to do
(provide a message, save, exit, or some combination of the above)
that might be useful as a command.

The default behaviour might go back to the current Exit behaviour for
new users, but not until I add an option to customize the behaviour
of the escape key, because I miss the old SaveAndExit behaviour.