Fixed bug where mouse was off by 1 char in tagline.
It turns out it was also a bug in the first line of the buffer.
The problem was that the default image map was advancing the start
of the rune rectangle on the first character of each new line. Since
the first line technically has no new line before it, the map was advanced
prematurely. It just wasn't noticable in the normal buffer because the
first line us usually "package x" and not modified, but the first line of
the tagline is the most used part of the tagline.

This fixes the logic up.

Closes #11