Added ability to view file as hex dump
This adds the ability to change the renderer being used to
view a file on the fly, and as a proof-of-concept adds a
hex editor renderer that you can switch to by running Renderer:hex
in any character buffer. You can't (or at least shouldn't) edit
the hex dump right now, since it doesn't show you where your cursor
is or have any way to see what you're doing, but it provides a way
to look at the exact bytes in the character buffer without getting
out of de.

All renderers must now register a name when they call RegisterRender
so that the Renderer:x command knows which renderer to use. Useful
renderers other than the default for a given file that can show any
file are either Renderer:hex for a hex dump or Renderer:nosyntax for
a renderer that doesn't bother with the niceties of syntax highlighting.