Refactored rendering code to minimize image allocations
There's no reason for Render to return an image which gets drawn,
when the screen buffer already has an *image.RGBA allocated. It's
more efficient to just draw directly into the screen buffer.

This refactors the renderer interface from having a
	Render(...) (image.Image, error)

method to having a
	RenderInto(dst draw.Image, ...) error

This results in minimal changes to memory usage, since the images
returned were short lived, but about 50% less CPU usage (from the
GC) when navigating around a lot. (ie. while holding a movement key
which triggers a lot of re-renders to move the cursor.) (Not measured
by any scientific benchmarks, but just by watching the resource usage
in top.)