Finally got text size to render properly on OS X (Both retina and Non)
While shiny correctly reports the PixelsPerPt in Shiny, OS X clearly
doesn't expect you to use it, because multiplying it through while
parsing fonts results in a constant physical size font on the screen
(as it should.) Meanwhile, OS X system preferences don't even refer
to resolution, they refer to "more text" and "less text" on the screen.

Unfortunately, the DPI that it expects you to parse fonts at is different
based on the screen. Retina displays expect it to be parsed at 144 DPI
to render correctly (where correctly is defined as "the same as other
non-shiny applications"), while normal displays expect it to be parsed at a
more standard 96 DPI.

Since we don't know what kind of display is being used, as a hack if it's
 > 120 DPI (half way between 96 and 144), we assume it's a retina display
and parse the font at 144 regardless of the reported resolution, otherwise
it's parsed at 96 DPI.