Added Undo buffer
The CharBuffer model now also maintains a pointer to the previous
buffer, acting as a singly-linked list of undos. This gets updated on
every insert or delete.

This could be implemented in a more memory-efficient way where it gets
serialized to disk and/or has a cap on the size of the undo buffer, but
for now, this is better than nothing and the memory usage is unnoticeable
most cases (ie. a source code file, even of a hundred kilobytes, being
duplicated in memory for a hundred edits before closing or opening another
file, is still only 10 megabytes on a machine which usually has gigabytes
of RAM.)