Fixed bug where viewport was moved too far before start of screen.
When focusing the viewport, it was blindly subtracting 100px from the
viewport in order to make the selected line appear a little farther
down the screen. This was causing a bug if you were on, for instance,
the first line where the viewport would be scrolled past the start of
the file.

This adds a check to make sure it only subtracts 100px if there is
100px to subtract from the viewport.