Added translucent warning mask for lines longer than 80 characters
This adds a subtle red translucent mask over the portion of the renderered
screen beyond 80 characters, so that you can tell where the terminal width
would end on a TTY. It's a very subtle mask (8/255) so that it's not overly
distracting, but enough to see when you're approaching it.

The mask can be configured with TermWidth:x to draw the mask after x characters
instead of 80. TermWidth:0 will turn off the mask, and TermWidth:-80 will invert
it so that the mask is drawn over the portion that *doesn't* exceed 80
characters. (I don't know why you would do that, it just seemed like the most
obvious meaning for a negative value.)

The strength of the alpha channel used for the mask can be configured with
WarnAlpha. WarnAlpha:1 will be nearly fully transparent, and WarnAlpha:255 will
be fully opaque. WarnAlpha:0 will reset the strength of the mask to its
default value.