Various minor bug fixes
The "fix `dd`" commit didn't work as well as advertised when there
was a repeat set.

Undo wasn't invalidating the image map cache, causing mouse clicks to register
incorrectly after an Undo which modified newlines.

'a' was acting weirdly at the end of a line after the aforementioned "fix `dd`"
commit.