Implemented Join / 'J' command to join multiple lines.
The executable word Join should work similarly to capital J in
vi (which should now work in de.) The difference between Join and J
is that J takes the number of lines to Join as the repeat modifier
(similar to vi) or joins the next 2 lines with no modifiers, while
Join takes the number of lines as an argument (ie. Join:3) to be
executed.

Both ignore their argument and Join all lines in the selected text
(as if in visual mode in vi) if something is selected.