Use cmd.exe instead of sh to spawn process on Windows
Middle clicking or pressing enter on an executable word currently
executes "sh" to interpret the command.

This works on Windows 10 where they have the whole Linux-On-Windows
thing going now, but doesn't work on older versions of Windows.

This updates the code to use cmd.exe instead of sh on Windows, because
it's more reliable and might as well use the native command.