Improve scroll wheel on Mac
This improves the scrolling experience on a Mac, by making each scroll
event a fixed amount instead of 1 line based on the font size (as is expected by
users of operating systems where the number of events is identical to the number
of clicks of the scroll wheel. OS X drivers send a variable number of events
depending on the speed that the scroll wheel was clicked, so changing it to
scroll by a fixed amount of the inverse number of pixels that the driver uses
to split the OS "scroll by x" pixels into n scroll events results in a user
experience that's more consistent with how other non-Go programs work.

Closes #12