Cache font metrics to avoid race condition
This caches all references to the font metrics that are used by renderers
in the renderer when the font face is calculated. The freetype library
is not thread safe, so this avoids code that only needs to look up a metric
without modifying anything from triggering the truetype code which loads
a glyph into memory, and potentially races with rendering code.