Added support for Plan9Ports plumber
Instead of poorly imitating the plumbing service, de can now
use the real p9p plumber on unix-like systems.

Since de itself is single-buffer, the new service deplumber must
be run running in order to receive plumbing messages and coordinate
whether a new de window should be created, or the existing one re-used.

If the plumbing message came from de and the buffer is clean, it will
tell that window to open the new file. Otherwise, it'll open a new one.

This frees de from having to do window management (leaving that job
to the window manager) while still getting the benefits of using
the real plumber.