Added ability to read from stdin and use as a pager
This has two related changes which improve the ability to use `de` in
Unix-like environments.

First, passing a filename of `-` will cause `de` to read the buffer from stdin
until it finds an EOF. This allows it to be used in pipeline chains such as
`ls | de -` (or, potentially more exotic pipes..)

Second, it adds the ability to use de as a standard pager. In pager mode it will
read from stdin without the `-` parameter instead of opening the current directory
when no filename is provided and it will default to the terminal emulation
renderer (Renderer:xterm) used by the Shell plugin. The terminal renderer has
also been improved to handle standard ANSI colour code and bold escape
sequences (though currently foreground colour only.)

Pager mode is detected by the command you use to invoke it. `de` will launch
as a regular text editor and `dep` will launch as a pager (you'll have to
manually copy the binary or create a link.) This is required in addition to
the pipe because some commands auto-page by writing to a temp file which is
passed as a parameter to $PAGER, and in this case `de` needs to know to change
renderers (if the command name is "less" or "more", pager mode will also be
detected.)

There should probably be command line arguments instead of magic based on the
command name, but there aren't any commands arguments at all in `de` right now,
and once some commandline arguments are added there are probably some that
would benefit from having different defaults based on the pager mode, which is
still easiest to detect by the command name.