Added hack to fork into background process.
Go doesn't make it very easy to fork without execing
something, so this adds a hack where de takes a --fg
command line parameter. If it isn't specified, it will
spawn a window with the fg parameter and exit.

This results in the terminal that de was spawned for
getting control back before the initial window quits.