Use a real PTY for the Shell command
This cleans up the shell plugin to create a PTY, rather than the
previous hack which tried to guess whether or not there was something
to read from the terminal and periodically sleep.

This is much more responsive and less CPU intensive, since the
plugin only reads when there's data to be read.

It also adds a mechanism to notify plugins when a mouse event occurs
(after the main routine has processed them), in order to allow the
Shell plugin to detect when a file has been opened via clicking
the mouse, so that it can kill the active shell process.