Handle plumbing messages on Plan 9.
This does a first pass of allowing deplumber to handle plumbing messages
on Plan 9, where the unix domain socket hack for coordinating the open
window doesn't work because there are no unix domain sockets.

This needs some work, and should create a proper 9fs server under /mnt/de,
rather than trying to port the hack, but for now it works, though it always
opens things in a new window.